adicionando validação de token no LaunchService e LaunchController e nos testes dele

+++ demo/src/main/java/com/money_track/demo/services/LaunchService.java
@@ -2,51 +2,83 @@ package com.money_track.demo.services;
 
 import com.money_track.demo.entities.DTO.LaunchDTO;
 import com.money_track.demo.entities.Launch;
+import com.money_track.demo.entities.User;
 import com.money_track.demo.repositories.LaunchRepository;
+import com.money_track.demo.repositories.UserRepository;
+import com.money_track.demo.security.TokenService;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class LaunchService {
-    private final LaunchRepository launchRepository;
 
-    public LaunchService(LaunchRepository launchRepository) {
+    private final LaunchRepository launchRepository;
+    private final TokenService tokenService;
+    private final UserRepository userRepository;
+
+    public LaunchService(LaunchRepository launchRepository, TokenService tokenService, UserRepository userRepository) {
         this.launchRepository = launchRepository;
+        this.tokenService = tokenService;
+        this.userRepository = userRepository;
     }
 
-    public List<LaunchDTO> findAllLaunches(){
-        List<Launch> allLaunches = launchRepository.findAll();
+    public List<LaunchDTO> findAllLaunches(String authHeader){
+        User user = findUserByToken(authHeader);
+        List<Launch> allLaunches = user.getLaunches();
         return allLaunches.stream().map(LaunchDTO::new).toList();
     }
 
-    public LaunchDTO findLaunchById(Long id){
+    public LaunchDTO findLaunchById(String authHeader,Long id){
+        User user = findUserByToken(authHeader);
         Launch launch = launchRepository.findById(id).orElseThrow(() -> new RuntimeException("erro"));
-        return new LaunchDTO(launch);
+        if(user.getLaunches().contains(launch)){
+            return new LaunchDTO(launch);
+        }else{
+            throw new RuntimeException("erro 2");
+        }
     }
 
 
-    public LaunchDTO createLaunch(Launch launch){
+    public LaunchDTO createLaunch(Launch launch,String authHeader){
+        User user = findUserByToken(authHeader);
+        launch.setUser(user);
         launchRepository.save(launch);
         return new LaunchDTO(launch);
     }
 
-    public LaunchDTO updateLaunch(Long id, Launch launch){
+    public LaunchDTO updateLaunch(Long id, Launch launch,String authHeader){
+        User user = findUserByToken(authHeader);
         Launch updatedLaunch = launchRepository.findById(id).orElseThrow(() -> new RuntimeException("erro"));
+        if(!user.getLaunches().contains(updatedLaunch)){
+            throw new RuntimeException("erro");
+        }
+        updatedLaunch.setUser(user);
         updatedLaunch.setCategory(launch.getCategory());
         updatedLaunch.setDate(launch.getDate());
         updatedLaunch.setDescription(launch.getDescription());
-        updatedLaunch.setUser(launch.getUser());
         updatedLaunch.setValue(launch.getValue());
         launchRepository.save(updatedLaunch);
         return new LaunchDTO(updatedLaunch);
     }
 
-    public void deleteLaunchById(Long id){
+    public void deleteLaunchById(Long id,String authHeader){
+        User user = findUserByToken(authHeader);
+        Launch deletedLaunch = launchRepository.findById(id).orElseThrow(() -> new RuntimeException("erro"));
+        if(!user.getLaunches().contains(deletedLaunch)){
+            throw new RuntimeException("erro");
+        }
         try{
             launchRepository.deleteById(id);
         }catch (RuntimeException e){
             throw new RuntimeException("erro");
         }
+    }
+
+    public User findUserByToken(String authHeader){
+        String token = authHeader.replace("Bearer ","");
+        String email = tokenService.validateToken(token);
+        return userRepository.findUserByEmail(email);
     }
 }
